@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>net5.0</TargetFramework>
+        <TargetFramework>net6.0</TargetFramework>
         <RuntimeIdentifiers>win-x64;win-x64;linux-x64;linux-arm</RuntimeIdentifiers>
         <GenerateDocumentationFile>true</GenerateDocumentationFile>
         <RootNamespace>WatcherBot.FSharp</RootNamespace>
