@@ -2,14 +2,14 @@
 
     <PropertyGroup>
         <OutputType>Exe</OutputType>
-        <TargetFramework>net5.0</TargetFramework>
+        <TargetFramework>net6.0</TargetFramework>
         <RuntimeIdentifiers>win-x64;win-x64;linux-x64;linux-arm</RuntimeIdentifiers>
         <Nullable>enable</Nullable>
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="DisCatSharp" Version="9.8.3"/>
-        <PackageReference Include="DisCatSharp.CommandsNext" Version="9.8.3"/>
+        <PackageReference Include="DisCatSharp" Version="9.8.4"/>
+        <PackageReference Include="DisCatSharp.CommandsNext" Version="9.8.4"/>
         <PackageReference Include="Microsoft.EntityFrameworkCore" Version="5.0.7"/>
         <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="5.0.7">
             <PrivateAssets>all</PrivateAssets>
