@@ -1,11 +1,10 @@
-﻿namespace WatcherBot
+﻿namespace WatcherBot;
+
+internal static class Program
 {
-    internal static class Program
+    public static void Main(string[] args)
     {
-        public static void Main(string[] args)
-        {
-            using var bm = new BotMain();
-            bm.MainAsync().GetAwaiter().GetResult();
-        }
+        using var bm = new BotMain();
+        bm.MainAsync().GetAwaiter().GetResult();
     }
 }
