@@ -3,12 +3,11 @@ using System.Threading.Tasks;
 using DisCatSharp.CommandsNext;
 using DisCatSharp.Entities;
 
-namespace WatcherBot.Utils
+namespace WatcherBot.Utils;
+
+[AttributeUsage(AttributeTargets.Method)]
+public class RequireDmOrOutputGuild : RequireOutputGuild
 {
-    [AttributeUsage(AttributeTargets.Method)]
-    public class RequireDmOrOutputGuild : RequireOutputGuild
-    {
-        public override async Task<bool> ExecuteCheckAsync(CommandContext ctx, bool help) =>
-            ctx.Channel is DiscordDmChannel || await base.ExecuteCheckAsync(ctx, help);
-    }
+    public override async Task<bool> ExecuteCheckAsync(CommandContext ctx, bool help) =>
+        ctx.Channel is DiscordDmChannel || await base.ExecuteCheckAsync(ctx, help);
 }
